refactor(ui): collocate pre-built schemas
