+++ packages/ui/package.json
@@ -28,6 +28,10 @@
       "import": "./dist/elements/media.js",
       "types": "./dist/elements/media.d.ts"
     },
+    "./schemas": {
+      "import": "./dist/schemas.js",
+      "types": "./dist/schemas.d.ts"
+    },
     "./typography": {
       "import": "./dist/elements/typography.js",
       "types": "./dist/elements/typography.d.ts"
