+++ packages/ui/src/elements/link.tsx
@@ -1,17 +1,8 @@
 import 'server-only'
 
-import type { Value } from '@deepdish/core/schema'
-import * as v from 'valibot'
 import { DeepDish } from '../deepdish'
 import type { ElementProps } from '../types'
-
-export const linkSchema = v.object({
-  destination: v.optional(v.string()),
-  href: v.optional(v.string()),
-  title: v.optional(v.string()),
-})
-
-type LinkValue = Value<typeof linkSchema>
+import type { LinkValue } from '../schemas'
 
 export function Link(props: ElementProps<'a', string>) {
   return (
